# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Universal AI Tools API
  description: |
    Comprehensive REST API for AI-powered applications using hybrid Go/Rust architecture.
    
    ## Features
    - JWT Authentication
    - Real-time Chat with Local LLMs (Ollama)
    - Vector Search with Qdrant
    - WebSocket Support
    - Agent Orchestration
    - Distributed Tracing
    
    ## Architecture
    - **Go**: API Gateway, WebSocket Service (high concurrency)
    - **Rust**: LLM Router, AI Core (high performance)
    - **TypeScript**: Legacy compatibility layer
    
  version: "1.0.0"
  contact:
    name: Universal AI Tools
    email: support@universal-ai-tools.com
    url: https://docs.universal-ai-tools.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8090
    description: Development server
  - url: https://api.universal-ai-tools.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/search:
    post:
      tags: [Vector Search]
      summary: Semantic search
      description: Perform semantic search on vector embeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Search query
                collection:
                  type: string
                  default: "universal_ai_docs"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 10
            example:
              query: "How does authentication work?"
              collection: "universal_ai_docs"
              limit: 5
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        score:
                          type: number
                        text:
                          type: string

  # Agent Management Endpoints
...
  /api/v1/conversation-context:
    get:
      tags: [Context]
      summary: Get context data
      description: Retrieve conversation context data with filtering options
      parameters:
        - name: category
          in: query
          schema:
            type: string
            description: Filter by context category
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Context data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContextEntry'

    post:
      tags: [Context]
      summary: Store context data
      description: Store new conversation context data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, content]
              properties:
                category:
                  type: string
                  description: Context category
                content:
                  type: string
                  description: Context content
                metadata:
                  type: object
                  description: Additional context metadata
                conversationId:
                  type: string
                  description: Associated conversation ID
      responses:
        '201':
          description: Context stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ContextEntry'
  /api/v1/conversation-context/search:
    get:
      tags: [Context]
      summary: Search context
      description: Search through stored context data
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            description: Search query
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContextEntry'
                  total:
                    type: integer

  # Hardware Authentication Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContextEntry:
      type: object
      properties:
        id:
          type: string
          description: Unique context entry identifier
        category:
          type: string
          description: Context category
        content:
          type: string
          description: Context content
        metadata:
          type: object
          description: Additional context metadata
        conversationId:
          type: string
          description: Associated conversation ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        relevanceScore:
          type: number
          description: Context relevance score

    # Hardware Authentication Schemas
...